requirements.txt
discord.py>=1.7.3
beautifulsoup4>=4.10.0
python-dotenv>=0.14.0
websockets>=8.1
sqlalchemy==1.4.46
